Normalizando membros para pessoas e membro não pode se remover
Dessa forma, o proprietário não pode se remover da atividade também

+++ app/Http/Controllers/PessoaController.php
@@ -11,30 +11,31 @@ use Illuminate\Support\Facades\Auth;
 
 class PessoaController extends DriveController
 {
-    public function salvarAdicionarPessoa(Request $request, $atividade_id){
+    public function salvarAdicionarPessoa(Request $request, $atividade_id)
+    {
         $pessoaAdicionada = User::where('email', '=', $request->input('email'))->get()->first();
-        if($pessoaAdicionada){
+        if ($pessoaAdicionada) {
             $usuarioLogado = User::find(Auth::id());
-            if($pessoaAdicionada->email != $usuarioLogado->email){
+            if ($pessoaAdicionada->email != $usuarioLogado->email) {
                 $atividadeAcademica = AtividadeAcademica::find($atividade_id);
                 $atividadeUsuarioExiste = AtividadeUsuario::where('user_id', '=', $pessoaAdicionada->id)->where('atividade_academica_id', '=', $atividadeAcademica->id)->get()->first();
 
-                if(!$atividadeAcademica->user_logado_proprietario()) {
+                if (!$atividadeAcademica->user_logado_proprietario()) {
                     return redirect()->back(); // Só o proprietario pode adicionar pessoas
                 }
-                
-                if(!$atividadeUsuarioExiste){
-                    if($atividadeAcademica){
+
+                if (!$atividadeUsuarioExiste) {
+                    if ($atividadeAcademica) {
                         $atividadeUsuario = new AtividadeUsuario();
                         $atividadeUsuario->user_id = $pessoaAdicionada->id;
                         $atividadeUsuario->atividade_academica_id = $atividadeAcademica->id;
                         $atividadeUsuario->save();
-    
+
                         $papelPessoaAdicionada = new Papel();
                         $papelPessoaAdicionada->nome = $request->input('papel');
                         $papelPessoaAdicionada->atividade_usuario_id = $atividadeUsuario->id;
                         $papelPessoaAdicionada->save();
-    
+
                         //Concedendo permissão no drive
                         $role = '';
                         if($papelPessoaAdicionada->nome == Papel::EDITOR){
@@ -56,28 +57,40 @@ class PessoaController extends DriveController
         return redirect()->back();
     }
 
-    public function editarPessoa(Request $request, $atividade_id) {
+    public function editarPessoa(Request $request, $atividade_id)
+    {
         $atividade = AtividadeAcademica::find($atividade_id);
-        if($atividade
-           && $atividade->user_logado_proprietario()
-           && in_array($request->papel, Papel::PAPEIS)) {
-                $atividade_usuario = AtividadeUsuario::find($request->atividade_usuario_id);
-                $atividade_usuario->papel->nome = $request->papel;
-                $atividade_usuario->papel->save();
+        if (
+            $atividade
+            && $atividade->user_logado_proprietario()
+            && in_array($request->papel, Papel::PAPEIS)
+        ) {
+            $atividade_usuario = AtividadeUsuario::find($request->atividade_usuario_id);
+            $atividade_usuario->papel->nome = $request->papel;
+            $atividade_usuario->papel->save();
         }
 
         return redirect()->back();
     }
 
-    public function removerPessoa(Request $request, $atividade_id) {
-         $atividade = AtividadeAcademica::find($atividade_id);
-        if($atividade
-           && $atividade->user_logado_proprietario()) {
+    public function removerPessoa(Request $request, $atividade_id)
+    {
+        $atividade = AtividadeAcademica::find($atividade_id);
+
+        if (
+            $atividade
+            && $atividade->user_logado_proprietario()
+        ) {
             $atividade_usuario = AtividadeUsuario::find($request->id_membro);
+
+            if ($atividade_usuario->user_id == Auth::id()) {
+                return redirect()->back();
+            }
+
             $atividade_usuario->papel->delete();
             $atividade_usuario->delete();
         }
 
-        return redirect()->back();       
+        return redirect()->back();
     }
 }
